Moving statistics to own tab
moving statistics to own tab and their own view controller
