+++ Open Access Helper/StatisticsController.swift
@@ -1,0 +1,58 @@
+//
+//  StatisticsController.swift
+//  Open Access Helper
+//
+//  Created by Claus Wolf on 31.01.20.
+//  Copyright © 2020 Claus Wolf. All rights reserved.
+//
+
+import Cocoa
+
+class StatisticsController: NSViewController {
+    
+    @IBOutlet weak var oaCount: NSTextField!
+    @IBOutlet weak var oaSearchCountLabel: NSTextField!
+    
+    
+    let preferences = Preferences()
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Implement this method to handle any initialization after your window controller's window has been loaded from its nib file.
+        updateCount()
+    }
+    
+    func updateCount(){
+        let count = readSettings(file: "count.txt")
+        let myOASearchCount = readSettings(file: "oacount.txt")
+        oaCount.stringValue = String(format: NSLocalizedString("So far we've helped you find %@ Open Access Documents!", comment: "shows on main window, number of OpenAccess found"), count)
+        
+        oaSearchCountLabel.stringValue = String(format: NSLocalizedString("You've had help with %@ Open Access searches", comment: "shows on main window, number of OpenAccess Searches conducted"), myOASearchCount)
+        
+    }
+    
+    func readSettings(file: String) -> String{
+        let file = file //this is the file. we will write to and read from it
+        
+        var text2 = ""
+        
+        let fileManager = FileManager.default
+        guard let groupURL = fileManager.containerURL(forSecurityApplicationGroupIdentifier: "J3PNNM2UXC.otzshare") else {
+            return ""
+        }
+        let safariExtDir = groupURL.appendingPathComponent("Library/Caches/")
+        let fileURL = safariExtDir.appendingPathComponent(file)
+        
+        //reading
+        do {
+            text2 = try String(contentsOf: fileURL, encoding: .utf8)
+        }
+        catch {
+            text2 = "0"
+        }
+        
+        
+        return text2
+    }
+    
+}
